requirements.txt
# Core packages
numpy==1.19.5
pandas==1.3.3
scikit-learn>=1.0.0
matplotlib==3.4.3
Pillow==8.3.1  # Modern fork of PIL
scipy>=1.7.0
tensorflow==2.7.0  # For deep learning models
argparse>=1.4.0    # For command-line arguments
h5py>=3.1.0        # For saving models in HDF5 format

# Optional but recommended packages
jupyter>=1.0.0
seaborn>=0.11.0
tqdm>=4.64.0       # Progress bars

# Additional packages
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pydantic==1.8.2
sqlalchemy>=1.4.0
python-dotenv